ユニオンレイド管理機能を追加し、Botの基本構造を実装
